use children for the link content of <PrefilledWorkflowLink />
