+++ plugins/scaffolder/src/components/PrefilledWorkflowLink/index.ts
@@ -13,4 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export { SerializedWorkflowLink } from './SerializedWorkflowLink';
+export { PrefilledWorkflowLink } from './PrefilledWorkflowLink';
